% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_AUCROC.R
\name{calc_AUCROC}
\alias{calc_AUCROC}
\title{Calculates area under the curve: receiver operating characteristic (AUC-ROC)}
\usage{
calc_AUCROC(scores, labels)
}
\arguments{
\item{scores}{vector with prediction scores}

\item{labels}{vector with class labels. Positive class: 1, negative class: 0}
}
\value{
AUC-ROC value
}
\description{
This function takes a scores vector and a labels vector and calculates AUC-ROC.
}
\examples{
s = c(1, 0.8, 0.6, 0.4, 0.2, 0)
l = c(1, 1, 0, 1, 0, 0)
calc_AUCROC(scores = s, labels = l)
[1] 0.8888889
}
