% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_kappa.R
\name{calc_kappa}
\alias{calc_kappa}
\title{Calculates Cohen's kappa}
\usage{
calc_kappa(tp, fp, fn, tn)
}
\arguments{
\item{tp}{Count of true positives: Integer}

\item{fp}{Count of false positives: Integer}

\item{fn}{Count of false negatives: Integer}

\item{tn}{Count of true negatives: Integer}
}
\value{
kappa value
}
\description{
This function takes true positive (tp), false positive (fp), false negative (fn), and true negative (tn) counts and calculates Cohen's kappa.
}
\examples{
calc_kappa(tp = 20, fp = 10, fn = 5, tn = 30)
[1] 0.5301205
}
