% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_MCC.R
\name{calc_MCC}
\alias{calc_MCC}
\title{Calculates Matthew's Correlation Coefficient (MSE)}
\usage{
calc_MCC(tp, fp, fn, tn)
}
\arguments{
\item{tp}{Count of true positives: Integer}

\item{fp}{Count of false positives: Integer}

\item{fn}{Count of false negatives: Integer}

\item{tn}{Count of true negatives: Integer}
}
\value{
MCC value
}
\description{
This function takes true positive (tp), false positive (fp), false negative (fn), and true negative (tn) counts and calculates MCC.
}
\examples{
calc_MCC(tp = 20, fp = 10, fn = 5, tn = 30)
}
