% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_median_error.R
\name{calc_median_error}
\alias{calc_median_error}
\title{Calculates median error between two vectors}
\usage{
calc_median_error(x1, x2)
}
\arguments{
\item{x1}{numeric vector 1}

\item{x2}{numeric vector 2}
}
\value{
median error value
}
\description{
This function takes two vectors and calculates their median error, i.e. median of the absolute difference. Vectors must be of equal length. Expected usage is that one vector is observed values and the second vector is predicted values.
}
\examples{
x <- runif(5)
y <- runif(5)
calc_median_error(x1 = x, x2 = y)
[1] 0.2433127
}
