% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.regression.R
\name{performance.regression}
\alias{performance.regression}
\title{Calculates performance metrics for a regression model}
\usage{
performance.regression(o, p)
}
\arguments{
\item{o}{numeric vector with observed values}

\item{p}{numeric vector with predicted values}
}
\value{
object with 7 performance metrics: $r, $r.p, $rho, $rho.p, $median_error, $MSE, $RSS
}
\description{
This function takes vectors of observed (o) and predicted (p) values and calculates 7 performance metrics: Pearson's r (+ P-value), Spearman's rho (+ P-value), median error, mean squared error, and residual sum of squares.
}
\examples{
x <- runif(5)
y <- runif(5)
performance.regression(o = x, p = y)
$r
cor
0.2370634

$r.p
[1] 0.701013

$rho
rho
0.4

$rho.p
[1] 0.5166667

$median_error
[1] 0.2336084

$MSE
[1] 0.5548035

$RSS
[1] 0.5548035
}
